1.1.Setting up with the Symfony Local Web Server Setup database in .env file. Email admin user is admin1@thespacebar.com, password is engage.
Run:
- composer install
- ./bin/console doctrine:database:create
- ./bin/console doctrine:migrations:migrate
- ./bin/console doctrine:fixtures:load

Downloading the Symfony Local Web Server

Find your browser and go to https://symfony.com/download. The Symfony local web server - or Symfony 'client' - is a single, standalone file that is full of superpowers. At the top, you'll see instructions about how to download it. These steps are different depending on your operating system - but it should auto-select the right one.

For me, I'll copy this curl command, find my terminal, paste and enter! This downloaded a single executable file called symfony. To make sure I can type that command from anywhere, I'll move this into a global bin directory. By the way, you only need to do these steps once on your computer... so you're done forever!

Unless we've mucked things up, we should now be able to run this from anywhere: try it!

symfony
Say hello to the Symfony CLI! It lists the most popular commands, but there are a lot more - run:
symfony help

Woh. We'll talk more about this tool in another tutorial. But, to start a local web server, just say:
symfony serve

Ah. The first time you run this, you'll get an error about running: symfony server:ca:install. Let's do that:
symfony server:ca:install

You'll probably need to type in your admin password. This command installs a local SSL certificate authority... which is awesome because when we run symfony serve, it creates a local web server that supports https! Woh! We get free https locally! Sweet!

Find your browser and go to https://127.0.0.1:8000 - or localhost, it's the same thing. Say hello to The SpaceBar! This is the app we've been building in our Symfony 4 series: a news site for space-traveling friends from across the galaxy.

Try logging in with admin1@thespacebar.com and password engage. Then go to /admin/article.

This is the admin section for the articles on the site. Each article has an image... but until now, that image has basically been hardcoded. Click to edit one of the articles. Our first goal is clear: add a file upload field to this form so we can upload the article image, and then render that on the frontend.

But we're going to keep things simple to start... and take a deep and wonderful look into the fundamentals of how files are uploaded on the web and how that looks inside Symfony. Let's go