3.1.Where & How to Store the File
For now, the form is still submitting to this test endpoint. We'll change that soon by moving it into the actual article form. But, to finish a successful file upload, we need move the uploaded file from the temporary spot on the filesystem to its final location.

Where to Store Uploads?

So... where should we store the uploaded article images? The first question to ask is: can these uploaded files be public to everyone? Or do we need to do some sort of security check before a user can view or download them? For article images, they can be public. But we'll talk about private files later.

Ok, so if someone needs to be able to view these images, it means they need to live somewhere in the public/ directory. Later, we're going to talk about storing files in the cloud! Like S3, which honestly, is an awesome idea. But right now, we're going to keep it simple and store things directly on our server.

So how about storing things in... I don't know... public/uploads? Create that new directory. Then, inside, create an empty .gitignore file. The reason I'm doing this might be confusing at first. My goal is to ignore any files added to this directory from git... because we don't want to commit uploaded files. But I would also like to make sure that this directory at least exists when I clone the repository.

Find your terminal and add the empty .gitignore file:
git add public/uploads/.gitignore

Next, open up the real .gitignore file - the one at the root of your app - and ignore the entire /public/uploads directory. It's a bit weird, but thanks to this, we will ignore all files in public/uploads except for the .gitignore file we already added.

Why did we do this? Well, unfortunately, you can't add a directory to git. So by adding this .gitignore file, it will guarantee that the public/uploads directory will exist when you clone the repository. Honestly, the file could be named anything, it's just sort of a common practice to use an empty .gitignore file for this.

Check it out: create a new file in public/uploads called foo. Then, find your terminal and run:
git status

We see the new public/uploads/.gitignore file but we do not see the foo file. That's perfect. Delete that.