3.2.Where & How to Store the File{Moving the Uploaded File}
Moving the Uploaded File

Let's get to work inside of our controller to move the file. First, set the uploaded file to a new $uploadedFile variable. And, unfortunately, the phpdoc on this get() method is a bit generic... so it doesn't tell our editor that this will be an UploadedFile object. Because I'm obsessed with auto-completion, let's add inline doc about this: this will be an UploadedFile object - but not the one from Guzzle - the one from HttpFoundation in Symfony.

And guess what? This UploadedFile object has a super useful method on it: move()! Give it the destination directory and it'll take care of the rest. To get that directory, say $destination = and we need to get the path to our uploads/ directory. The best way is to read a parameter: $this->getParameter('kernel.project_dir') - to get the absolute path to the root of the app - then /public/uploads. Then add $uploadedFile->move() and pass it .

Hold Command or Ctrl and click this method. Ah, it returns a File object that represents the new file. Let's see what this looks like: surround this entire call with dd().

Alright team! Find your browser, refresh and re-post that upload. I... think it worked! The dumped file object tells me that there is a new file in our public/uploads/ directory. Let's go check it out! There it is! Well, I think that's it... but sheesh - the filename is terrible. Let's check its file size:
ls -la public/uploads/

Yea... that looks correct - it's about 1.8 megabytes. So... we moved the file... but that is a terrible filename. Let's fix that next.