6.1.Centralizing Upload Logic{Creating the Service}
We've got a pretty nice system so far: moving the file, unique filenames and putting the filename string into the database. But it is kind of a lot of logic to put in the controller... and we already need to reuse this code somewhere else: in the new() action.

Creating the Service

That's why I like to isolate my upload logic into a service class. In the Service/ directory - or really anywhere - create a new class: how about UploaderHelper?

This class will handle all things related to uploading files. Create a public function uploadArticleImage(): it will take the UploadedFile as an argument - remember the one from HttpFoundation - and return a string. That will be the string filename that was ultimately saved.

Ok! Let's go steal some code for this. In fact, we're going to steal pretty much all the logic here... and paste it in. Make sure to retype the r on Urlizer to get the use statement on top.
And at the bottom, return $newFilename.

Perfect! Well... not perfect, because the $this->getParameter() method is a shortcut that only works in the controller.