6.2.Centralizing Upload Logic{Creating the Service}
If you need a parameter - or any configuration - from inside a service, you need to add it via dependency injection. Add the public function __construct() with, how about, a string argument. Instead of just injecting the kernel.project_dir parameter, we'll pass in the whole string to where uploads should be stored.

I'll put my cursor on that argument name, hit Alt + Enter and select initialize fields to create that property and set it. Now, below, we can say $this->uploadsPath and then /article_image.

Cool! Let's worry about configuring the $uploadsPath argument to our service in a minute. After all, Symfony's service system is so awesome, it'll tell me exactly what I need to configure once we try this.