6.3.Centralizing Upload Logic{Creating the Service}
For now, go back into ArticleAdminController and use this. Start by adding another argument: UploaderHelper $uploaderHelper. And celebrate by removing all of the logic below and replacing it with $newFilename = $uploaderHelper->uploadArticleImage($uploadedFile).

Dang - that is nice! There is still a little bit of logic here: the form logic and the logic that sets the filename on the Article - but I'm comfortable with that. And we now have this great new method: pass it an UploadedFile object, and it'll move it into the correct directory and give it a unique filename.

Let's take it for a test drive! Go back, refresh the form and... it works! Naw, I'm kidding - we knew this error was coming... but isn't it beautiful?

Cannot resolve argument $uploadHelper of the edit() method: Cannot autowire service UploadHelper: argument $uploadsPath of method __construct() is type-hinted string, you should configure its value explicitly.

That's programming poetry people! And it makes sense: autowiring doesn't work for scalar arguments.