6.4.Centralizing Upload Logic{Creating the Service}
Open config/services.yaml. We could configure the specific argument for this specific service. But if you've watched our Symfony series, you know that I like to use the bind feature. The argument name is $uploadsPath. So, below _defaults and bind, add $uploadsPath set to %kernel.project_dir%/public/uploads.

This means: anywhere that $uploadsPath is used as an argument for a method that's autowired - usually a controller action or the constructor of a service - pass in this value.

Exceeding upload_max_filesize

Let's go see if that fixed things - reload. Now we see the form. To test this fully, let's empty out the article_image/ directory. This time, let's upload the stars photo. Hit update.

Why is this happening? I'll give you a clue: the file we just uploaded is 3mb. Go to your terminal and run
php -i | grep upload