6.5.Centralizing Upload Logic{Adding the Logic to new() Action}
Adding the Logic to new() Action

Now that all of our logic is isolated, we can easily repeat this in the new() action. We do need to copy these 5 lines or so, but I'm happy with that.

Up in new(), add the argument - UploaderHelper $uploaderHelper - and inside the isValid() block, paste!

This uses the same form, with the same unmapped field, so it'll all just work.

Next: let's talk about validation.