9.5.URL to Public Assets{Using a Service Subscriber}
Using a Service Subscriber

Inside: we need to get the UploaderHelper service so we can call getPublicPath() on it. Normally we do this by adding it as an argument to the constructor. But, in a few places in Symfony, for performance purposes, we should do something slightly different: we use what's called a 'service subscriber', because it allows us to fetch the services lazily. If this is a new concept for you, go check out our Symfony Fundamentals course(https://symfonycasts.com/screencast/symfony-doctrine/service-subscriber) - it's a really cool feature.

The short explanation is that this class has a getSubscribedServices() method where we can choose which services we need. These are then included in the $container object and we can fetch them out by saying $this->container->get().

Add UploaderHelper::class to the array.

Then, above, we can return $this->container->get(UploaderHelper::class)->getPublicPath($path).

Let's give it a try! Refresh! We got it! That took some work, but I promise you'll be super happy you did this.

Next: let's also update the image path in the show page, and learn a bit about what the asset() function does internally and how we can do the same thing automatically in UploaderHelper.